+++ plugins/extensions/plugins/ZeroEQPlugin.h
@@ -32,6 +32,7 @@
 #include <lexis/render/imageJPEG.h>
 #include <lexis/render/lookupTable1D.h>
 #include <lexis/render/viewport.h>
+#include <lexis/render/histogram.h>
 #include <zerobuf/render/attribute.h>
 #include <zerobuf/render/colormap.h>
 #include <zerobuf/render/frameBuffers.h>
@@ -208,6 +209,13 @@ private:
     void _viewportUpdated();
 
     /**
+     * @brief This method is called when the histogram for the current simulation frame is requested
+     *        by a ZeroEQ event
+     * @return True if the method was successful, false otherwise
+     */
+    bool _requestHistogram();
+
+    /**
      * @brief Resizes an given image according to the new size
      * @param srcData Source buffer
      * @param srcSize Source size
@@ -247,6 +255,7 @@ private:
     ::lexis::render::ImageJPEG _remoteImageJPEG;
     ::lexis::render::LookupTable1D _remoteLookupTable1D;
     ::lexis::render::Viewport _remoteViewport;
+    ::lexis::render::Histogram _remoteHistogram;
 
     ::brayns::v1::DataSource _remoteDataSource;
     ::brayns::v1::Settings _remoteSettings;